E26 compatibility

+++ usr/lib/user-manager/hooks.d/e17-wizard/last/90-unset-gtk-theme.sh
@@ -1,4 +1,6 @@
 #!/bin/sh
 
 # wizard is finished, we don't need a default gtk settings anymore
-killall xsettingsd 2>/dev/null 1>&2 || true
+if [ -n "$EROOT" ] ; then
+    killall xsettingsd 2>/dev/null 1>&2 || true
+fi
